feat(fron-kit): moved unit test for dynamicroute
