+++ packages/falcon-front-kit/package.json
@@ -15,11 +15,18 @@
     "dist",
     "src"
   ],
+  "jest": {
+    "setupFilesAfterEnv": [
+      "./../../test/setupTests.js"
+    ]
+  },
   "scripts": {
     "prepare": "npm run clean && npm run build",
     "clean": "falcon-scripts clean",
     "build": "falcon-scripts build",
-    "watch": "falcon-scripts watch"
+    "watch": "falcon-scripts watch",
+    "test": "falcon-scripts test",
+    "test:coverage": "falcon-scripts test:coverage"
   },
   "dependencies": {
     "@deity/falcon-data": "^0.0.1",
@@ -31,8 +38,11 @@
   "devDependencies": {
     "@deity/falcon-scripts": "^0.0.1",
     "@types/react-router-dom": "4.3.4",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.7.0",
     "prop-types": "15.7.2",
     "react": "^16.6.3",
+    "react-apollo": "2.5.1",
     "react-router-dom": "5.0.1"
   },
   "peerDependencies": {
